refactor(coquille): `app.js` disparaît — il s'appelle `amorce.js` et ne fait plus qu'amorcer
L'ADR-011 se termine. `app.js` — 2 869 lignes il y a un mois — est renommé `amorce.js` et n'en
porte plus que 497, dont l'essentiel est une SÉQUENCE : brancher, monter, écouter, charger,
restaurer, peindre. `index.html` et `sw.js` le nomment sous son nouveau nom.

Six choses en sortent, chacune vers la maison qui lui va :

— **le CYCLE de rendu monte dans `rendu.ts`.** Ce module était un CROCHET, et son propre en-tête
  annonçait sa fin : « transitoire, il disparaît quand la carte du voyage, la soute et les entrepôts
  auront emménagé dans l'arbre ». Ils y sont. `brancherRendu` n'existe plus, `rafraichir()` EST le
  cycle, et `rafraichirDifferee` y est UNE instance partagée — six champs se partagent son timer, et
  chaque module qui en referait un ferait six rendus là où il y en avait un ;
— **la NAVIGATION dans `navigation.js`.** Les vingt-trois `hidden = v !== "…"` deviennent une table
  vue → sections, les huit `classList.toggle` une boucle sur `data-view`, et les huit écouteurs du
  rail UNE délégation. Les touches 1…8 se dérivent désormais de l'ordre du rail : le numéro lu sur
  un bouton EST la touche qui l'ouvre, par construction et non plus par discipline ;
— **le TRI dans `tri.js`**, idempotent (il ne l'était pas : un second appel doublait l'inversion) ;
— **`syncToggles` dans `filtres.ts`** sous le nom `synchroniserReglages` — même quatorze contrôles
  que `readFilters`, même doctrine ;
— **les relevés d'autoload dans `frais-actions.ts`**, miroir de `corrections-actions.ts` ;
— **les quatre gestes des formulaires de soute dans `soute-actions.js`**, avec leurs deux
  `flushSync` : `amorce.js` n'en contient plus aucun, le dépôt en compte toujours quatre.

Deux ancres textuelles sur `app.js` sont réancrées, et les deux y gagnent :

— `logic.test.mjs` gardait `fiabiliteCell`, du code MORT dont ce test était le seul appelant. Il lit
  maintenant `CelluleFiabilite` (vues/communs.tsx), le rendu qui s'affiche vraiment. Vu ROUGE sur le
  vrai composant avant que le jumeau ne parte ;
— `scripts/jetons.test.mjs` lisait `app.js` pour les jetons cités depuis le JavaScript ; il aurait
  levé ENOENT. Il lit `vues/carte.tsx`, seul fichier de code du dépôt qui écrive `var(--…)`. Cette
  seconde lecture est INERTE aujourd'hui — mesuré, les quatre jetons concernés ont tous un usage
  CSS — et on la garde pour le jour où ce ne sera plus vrai.

Au passage : la génération du compagnon passe de `voyage-actions.ts` à `voyage-donnees.ts` (un
compteur n'est pas une action, et le cycle doit l'atteindre sans cycle d'import), 39 imports morts
partent d'`amorce.js`, et `chargerJambe` cesse d'appeler `rafraichir()` deux fois de suite.

Vérification : `node --test` 490/490 ; `npx playwright test` 244 tests, 242 passés, 2 ignorés,
0 échec ; `npx tsc --noEmit` muet ; `npm run build:site` OK, précache 20 entrées.
