--- app.js
@@ -1,1334 +1,0 @@
-"use strict";
-
-// Fonctions de calcul pures (testées par logic.test.mjs).
-import {
-  ageDays, pairAge, scoreBarWidth, addableUnits, cargoBoxes, kFromReading, kPlausible, ovKey, safeKey, routeMetrics, enRouteDeals, bestManifest, palierMarge, freeManifestLine, hydrateManifestLine, stationLabel, stationTree, manifestJourneyState, offloadPlan,
-} from "./logic.ts";
-// Le premier îlot React (ADR-008, #96). `peindre` remplace `innerHTML` sur le conteneur de la
-// Tournée, et RIEN d'autre ne change : app.js reste le seul écrivain de l'état, refresh() reste
-// l'unique notification. Voir pont.js pour pourquoi il n'y a ni magasin ni abonnement.
-import { flushSync } from "react-dom";
-import { etat, notifier } from "./etat.ts";
-import { esc, fmt, signe } from "./format.ts";
-import { readFilters } from "./filtres.ts";
-import {
-  effVals, isOv, loadOverrides, ovCount, resetOverrides, saveOverrides, setOverride,
-} from "./corrections.ts";
-import { corriger, notifySuperseded, updateOvBadge } from "./corrections-actions.ts";
-import { showToast } from "./messages.ts";
-import {
-  construireIndex, indexOrigine, indexStationExacte, libellesOrigines, libellesStations, resolveStationLabel, stationCourante, stationMap, termByName,
-} from "./marche.ts";
-import { alKey, kFmt, loadAutoloadK, saveAutoloadK } from "./frais.ts";
-import { applyState, loadState, saveState } from "./persistance.ts";
-import { brancher, withMarket } from "./donnees.ts";
-import { brancherRendu, rafraichir } from "./rendu.ts";
-import {
-  addManifestCommodity, addSuggestion, oublierCompositionSiRouteChangee, removeManifestLine, resetManifeste, updateManifestTotals,
-} from "./manifeste-gestes.js";
-import {
-  copierCorrections, copierEntrepots, copierPlan, copyManifest, copyShareLink,
-} from "./presse-papiers.js";
-import {
-  addLegLine, addLegSuggestion, addStopByTerminal, beginJourney, clearJourney, delLegLine, editLegQty, liveLegQty, manifestToJourney, removeJourneyStop, resetLeg, setJourneyStop, toggleLegEditor,
-} from "./voyage-gestes.js";
-import {
-  chargerJambe, declarerABord, deposerIci, loadChargements, loadDepots, loadSoute, poserPosition, reprendreIci, retirerLot, vendreIci, venteImplicite, viderSoute,
-} from "./soute-actions.js";
-import { manifestRemaining, suggestionsFor } from "./manifeste-donnees.ts";
-import {
-  compositionValide, loadManifestEdit, nouvelleGenerationManifeste, oublierComposition,
-} from "./manifeste-etat.ts";
-
-
-import { nouvelleGenerationVoyage, pickJourney, syncViewsToJourney } from "./voyage-actions.ts";
-import {
-  chargerVaisseaux, memoriserStation, monterSelecteurStation, montrerCarteVaisseau, stationChangee,
-} from "./selecteur.js";
-import { monterRacine } from "./main.tsx";
-import { planData } from "./vues/plan-vue.tsx";
-import {
-  figerJambe, legSuggestCtx, journeyCarriedCommodities, legIntent, loadJourneyEdits, loadJourneyPins, pinLegsForVolume,
-} from "./voyage-donnees.ts";
-// (`vues/tournee.tsx` et `vues/plan.tsx` ne sont plus importés ici : leurs vues vivent dans l'arbre
-// depuis #143 et #145, et seuls leurs composants de DÉCISION les consomment désormais.)
-// La vue Commodités n'expose plus sa présentation à app.js — seulement ses trois ACTIONS, comme
-// `plan-vue.tsx` expose `planData`. Les écouteurs de `#commSortModes` / `#commBoardModes` restent
-// ici : leurs conteneurs sont du markup d'index.html (ADR-012 §2).
-import { refletBoardCommodites, setCommBoard, setCommSort } from "./vues/commodites-vue.tsx";
-
-
-
-
-
-import { BUY_STATUS, KIND_ICON, SELL_STATUS } from "./vues/communs.tsx";
-
-// Libellé compact des caisses SCU standard, ex. « 8×32 · 1×16 · 1×4 · 1×2 · 1×1 ».
-// `maxBox` = plafond de caisse du terminal de CHARGEMENT, quand on le connaît : c'est une propriété
-// physique de la station, indépendante de l'interrupteur de frais. On le propage partout où le
-// terminal d'achat est disponible, parce que c'est exactement la décomposition que la facture
-// d'autoload utilise — un « 📦 1×32 » à côté d'un montant calculé sur deux caisses de 16 serait
-// une incohérence directement visible.
-
-// Même libellé pour un chargement à plusieurs commodités : une caisse ne contient qu'une commodité,
-// la décomposition se fait donc ligne par ligne (cargoBoxes) et jamais sur le total des SCU.
-
-// État global
-// Tri par défaut : le PROFIT NET par voyage (ADR-005). Le score composite classait mal — la route
-// la plus rentable de l'instantané tombait au 8e rang — et le profit horaire repose sur une durée
-// fictive pour 49 % des routes, faute de distance. Un montant, lui, ne ment pas.
-// Les CINQ caches de rendu du board Commodités — `shownCommodities`, `commTiers`, `commDupCodes`,
-// `commMaxMargin`, `commCarried` — sont devenus des variables locales de `vues/commodites-vue.tsx`.
-// Ils n'existaient que parce qu'une fonction calculait et que trois autres peignaient ; un composant
-// qui fait les deux dans la même passe n'en a pas besoin. Aucun n'avait de lecteur hors de la vue.
-// Compagnon de voyage : parcours sélectionné { legs[], current } ou null.
-
-const $ = (id) => document.getElementById(id);
-// Volume dont le null veut dire « capacité non communiquée par UEX » et non « zéro » :
-// `scu_sell` n'est renseigné que sur une minorité de points de vente. Un « — » s'y lisait
-// « aucune demande » alors qu'aucun plafond n'est appliqué dans ce cas — d'où « n.c. ».
-
-// L'échappement HTML (`esc`) et le badge de système sont partis : le premier dans `format.ts`, où il
-// a une seule source pour tous ses appelants, le second dans `selecteur.js`, son unique lecteur.
-
-
-// L'icône de commodité et le marqueur « illégal » sont rendus par `IconeCommodite` et `TagIllegal`
-// (vues/communs.tsx). Leurs versions en chaîne ont disparu avec `multiCargoHTML`, leur dernier
-// appelant. `KIND_ICON` reste importé : la table d'emoji n'a qu'une source.
-
-// ---------- Fiabilité : fraîcheur, statut de stock, aberrations ----------
-// (ageDays/pairAge et les calculs de temps/score viennent de logic.mjs)
-// Fraîcheur d'une ligne de manifeste = le plus ancien des relevés achat/vente (ou l'un des deux).
-
-
-// `BUY_STATUS` et `SELL_STATUS` sont passées dans `vues/communs.tsx`, à côté de la pastille qui les
-// lit : app.js ne faisait que les repasser en props à deux vues.
-
-// ---------- Corrections locales (prix & stock) ----------
-// L'utilisateur peut corriger un prix ou un volume (stock à l'achat / demande à la vente)
-// quand le relevé UEX est faux. Stocké UNIQUEMENT en local (localStorage), jamais partagé
-// ni dans l'URL. Clé : « commodité|terminal|side » (side = "buy" | "vol"… non : "buy"/"sell").
-const nowSec = () => Math.floor(Date.now() / 1000);
-
-// ---------- Frais d'autoload : tarif par station + contexte de calcul ----------
-// Le calcul est PUR et vit dans logic.mjs (autoloadFee / autoloadPoint / haulFee). app.js n'y
-// apporte que ce que logic.mjs ne peut pas deviner sans lire une globale : quel terminal porte
-// quel nom, et combien CETTE station facture. Deux résolutions, donc :
-//   - nom de terminal -> terminal de market.json, parce que routes.json et loops.json (vues
-//     « Trajets » et « Boucles ») ne portent que des noms ;
-//   - terminal -> coefficient `k`, relevé par l'utilisateur ou valeur globale par défaut.
-// `showToast` est passée dans `messages.ts`, `notifySuperseded` dans `corrections-actions.ts` :
-// toutes deux doivent être appelables depuis une vue de l'arbre (ADR-012).
-
-// Applique les corrections à une paire buy/sell et renvoie des copies patchées + marge/roi.
-
-
-// Calcule les champs dérivés d'une route selon les entrées utilisateur : applique les corrections
-// locales (impur, globales OVERRIDES) puis délègue le calcul pur à routeMetrics (logic.mjs).
-
-
-// Cellule de FIABILITÉ : mini-barre + valeur, de 10 à 100 (ADR-005). Ce n'est plus un classement
-// mais ce qu'on sait de la donnée — fraîcheur du relevé × part de volume publiée. Elle ne trie plus
-// par défaut : c'est le profit net qui le fait. `scoreBarWidth` borne le dessin par ceinture (#39),
-// même si la fiabilité ne peut plus sortir de [0, 100] par construction.
-//
-// NE PAS SUPPRIMER, malgré l'absence d'appelant : le rendu vit dans `CelluleFiabilite`
-// (vues/communs.tsx), mais `logic.test.mjs:117` lit CE fichier comme du TEXTE et cherche
-// `function fiabiliteCell(` pour vérifier que la largeur passe bien par `scoreBarWidth`. La retirer
-// fait tomber un test unitaire, et l'ancre est délibérée — c'est le seul moyen qu'a un test de
-// `node --test` de garder une règle qui vit dans app.js.
-function fiabiliteCell(f, age, part) {
-  const tier = f >= 70 ? "s-good" : f >= 40 ? "s-ok" : "s-low";
-  const quoi = age == null ? "date du relevé inconnue" : `relevé vieux de ${Math.round(age)} j`;
-  const titre = `Fiabilité ${f}/100 — ${quoi}, ${Math.round(part * 100)} % du volume publié par UEX. N'entre pas dans le tri.`;
-  return `<div class="score-cell" title="${esc(titre)}"><span class="scorebar ${tier}"><i style="width:${scoreBarWidth(f)}%"></i></span><b>${f}</b></div>`;
-}
-
-// La valeur éditable est rendue par `ValeurEditable` (vues/communs.tsx). `editv()` fabriquait le
-// même span en chaîne pour les vues restées impératives ; la dernière — le manifeste — est passée
-// à React, donc plus personne ne l'appelait. `e2e/edition.pw.mjs` vérifie qu'aucun `.editv` de
-// l'application n'échappe à React, ce qui est la condition de cette suppression.
-
-// Lit l'état de tous les contrôles de filtre (partagé par les deux vues).
-
-
-// Message de #empty tel qu'il est écrit dans index.html. Le <p> est PARTAGÉ par les vues Trajets /
-// Boucles / En route, et « En route » comme le mode multi-commodité réécrivent son texte : sans
-// remise à zéro en tête de rendu, un état vide légitime affichait le message d'une AUTRE vue.
-
-// Ce que les deux modes de Trajets partagent. app.js garde l'ÉTAT — les frais dépendent de
-// l'interrupteur d'autoload et des relevés par station, l'écriture d'une correction doit figer les
-// jambes déjà planifiées, et `#empty` reste écrit ici (il est partagé par trois vues).
-
-
-// Ce qui se calcule LIGNE PAR LIGNE, et qui vaut pour les deux tables à lignes simples : « Trajets »
-// (`#rows`) et « En route » (`#enrouteRows`). Elles ont toujours partagé leur rendu — c'était la
-// fonction `routeRowHTML`, appelée des deux endroits. L'îlot React la remplace, donc il doit servir
-// les deux appelants : ne peindre que `#rows` laisserait « En route » sans lignes.
-
-
-
-
-// ---------- Vue « Trajets » en mode MULTI-COMMODITÉ ----------
-// Même tableau, mais chaque ligne est un chargement A->B composé de PLUSIEURS commodités
-// (remplissage par marge décroissante, plafonné par stock/demande et budget).
-
-
-// Le chargement déplié d'un trajet multi est rendu par `ChargementDeplie` (vues/trajets.tsx), avec
-// l'état d'ouverture — `multiCargoHTML` y a disparu, et `commodityIcon`/`illegalTag` avec lui : ils
-// n'avaient pas d'autre appelant.
-
-// Ligne de tableau pour une route évaluée (partagée par « Trajets simples » et « En route »).
-// `routeRowHTML` a été remplacé par vues/trajets.tsx, pour `#rows` ET `#enrouteRows`.
-
-
-
-
-
-
-// ---------- Mode « En route » (trajet dirigé) + manifeste multi-commodité ----------
-let enrouteReady = false;     // datalist/destSystem peuplés une seule fois
-// Nom de terminal -> terminal de market.json. Pont indispensable aux frais d'autoload : routes.json
-// et loops.json ne portent QUE des noms, et les noms sont déjà la clé métier du dépôt (corrections
-// locales, jambes de voyage). Peuplée en même temps que stationMap.
-// Le garde du champ #station (« ne pas re-rendre si la station n'a pas changé ») vit avec le
-// sélecteur qui l'écrit : `memoriserStation` / `stationChangee` (selecteur.js).
-
-// Charge le graphe de marché à la demande. Deux règles, apprises à la dure :
-//   - on mémorise la PROMESSE en vol, pas seulement son résultat : sinon chaque frappe pendant le
-//     chargement relançait un fetch complet de market.json (4 requêtes concurrentes mesurées) ;
-//   - on ne mémorise JAMAIS l'échec. Un marché vide mis en cache verrouillait « En route »,
-//     « Chaîne », « Commodités » et « Corrections » pour TOUTE la session — autocomplétion vide,
-//     0 tuile, « aucune chaîne rentable » — sans le moindre message, et seul un rechargement
-//     complet réparait. L'erreur remonte donc aux appelants, et l'action suivante réessaie.
-// Prévient que le marché est indisponible plutôt que de laisser la vue vide ET muette.
-const marketUnavailable = () => showToast("⚠ Marché indisponible — vérifie ta connexion, puis réessaie");
-
-// Peuple la liste des terminaux de départ (ceux où l'on peut acheter). Idempotent.
-function setupEnRoute() {
-  if (enrouteReady) return;
-  // Les trois index viennent de `marche.ts` ; ici on ne fait plus que peindre les listes.
-  construireIndex(etat.MARKET);
-  $("originList").innerHTML = libellesOrigines().map((l) => `<option value="${esc(l)}"></option>`).join("");
-  // Datalist de TOUTES les stations (achat ou vente) pour la vue Corrections.
-  $("stationList").innerHTML = libellesStations().map((l) => `<option value="${esc(l)}"></option>`).join("");
-
-  // Datalist de TOUTES les commodités (pour l'ajout libre au manifeste).
-  $("commodityList").innerHTML = etat.MARKET.commodities
-    .map((c) => `<option value="${esc(c.name)}">${esc(c.code || "")}</option>`).join("");
-
-  // Le sélecteur de station (ADR-003) se monte ICI et une seule fois : il lui faut MARKET, et
-  // `setupEnRoute` est le seul point du cycle où le marché est garanti présent.
-  monterSelecteurStation();
-
-  enrouteReady = true;
-}
-
-// Résout le terminal de départ depuis le texte du champ (libellé exact).
-
-// Résout le terminal d'ARRIVÉE forcé (En route) depuis le champ (libellé exact), ou null.
-
-
-// dealFrom / enRouteDeals / bestManifest / buildChainAdjacency vivent dans logic.mjs (fonctions
-// pures) ; on leur passe MARKET et le résolveur de corrections effVals depuis les vues.
-
-// Le chargement courant se DÉRIVE à la demande — plus de globale écrite au rendu et relue au clic.
-// C'est la même leçon qu'`indexOrigine` : une valeur dérivée qu'il faut penser à recalculer est une
-// valeur qui sera un jour lue périmée. Le coût est un `bestManifest` par geste — celui-là même que
-// le rendu que le geste déclenche referait de toute façon.
-
-// ---------- Le chargement qu'on COMPOSE à la main (#19) ----------
-// Lignes ajoutées, SCU ramenés à ce qu'on veut vraiment acheter : cette intention n'existait que
-// dans une globale que le rendu remettait à null à chaque passe. Un prix corrigé, une
-// frappe dans la recherche ou un vaisseau changé l'effaçaient sans un mot. Elle se persiste donc
-// comme le manifeste d'une jambe de voyage, et SOUS LA MÊME FORME : l'INTENTION seule —
-// [{ name, units }] — jamais un instantané de marché, dont les prix se figeraient au jour de
-// l'édition (cf. hydrateManifestLine). Le couple de terminaux l'accompagne : c'est de LEURS prix
-// que ses lignes sont relues, et une composition ne vaut que pour la route qu'on composait.
-// Une seule à la fois, là où JOURNEY_EDITS en indexe une par jambe : la carte n'en montre qu'une, et
-// une composition rangée sous un couple qu'on ne regarde plus ressortirait des jours plus tard sans
-// qu'on l'ait demandée. Le système est stocké avec le nom — deux terminaux peuvent porter le même
-// nom dans deux systèmes, et les lignes se relisent par INDEX de terminal.
-
-
-// Retient ce qui est à l'écran comme intention. Appelée par CHAQUE geste de composition — ajout
-// suggéré, ajout libre, retrait, SCU ajustés — parce que c'est le geste qui fait la composition,
-// pas son résultat : deux gestes qui se compensent laissent quand même une carte à soi.
-
-// « ↺ optimal » : la carte redevient un calcul, et se remet à suivre le marché et les filtres.
-
-/**
- * Efface la composition si la route affichée n'est plus la sienne.
- *
- * C'est l'ancienne décision de `compositionEnCours`, sortie du rendu. On la rejoue au GESTE, avec
- * la même règle et le même juge (`compositionValide`) : ce qui change, c'est QUI décide — un clic
- * de l'utilisateur, et non un repaint provoqué par une correction de prix ailleurs.
- */
-
-
-// La marque et le contrôle du chargement composé à la main. Les deux mêmes chaînes servent au rendu
-// de la carte ET à leur pose en direct à la première frappe dans un champ SCU : celle-là ne repeint
-// pas la carte (elle arracherait le champ qu'on remplit), et sans ces deux-là rien à l'écran ne
-// dirait que la carte est désormais la tienne, ni comment la rendre au calcul.
-// Le ✎ et le bouton « ↺ optimal » sont rendus par vues/manifeste.tsx d'après `MANIFEST_EDIT`.
-// Ils étaient injectés par `insertAdjacentHTML` DANS une carte déjà rendue, faute de pouvoir la
-// repeindre sans arracher le champ en cours de frappe — la contrainte a disparu avec l'îlot.
-
-// La composition vaut-elle pour la carte demandée ? Sinon elle est abandonnée SUR PLACE : la garder
-// en réserve la ferait ressurgir au retour sur cette route, longtemps après le geste qui l'a écrite.
-// Rend { edit, destIdx } ou null ; `destIdx` est l'arrivée que la carte doit alors afficher.
-
-
-// Lignes d'une composition, RELUES au marché courant : c'est ce qui fait qu'un prix corrigé
-// s'affiche alors que les SCU, eux, ne bougent pas.
-
-// Carte d'un couple de terminaux dont plus AUCUN chargement n'est rentable : bestManifest ne rend
-// alors rien, et la composition faite à la main disparaîtrait avec lui — alors que le geste qui
-// vient de tuer la rentabilité (un prix d'achat corrigé vers le haut) est justement celui qui la
-// rend précieuse. Mêmes champs que ceux que manifestsFrom estampille sur un trajet ; les lignes
-// viennent de l'appelant.
-
-
-// `isOv` est passée dans `corrections.ts`, à côté d'`effVals` : elle lit le même store.
-
-// `m` = manifeste courant (il porte `fee`, le contexte de frais qui l'a produit) ; `t` = ses totaux.
-// Les totaux du manifeste sont rendus par vues/manifeste.tsx (`<Totaux>`).
-
-// Espace/budget restants d'après les SCU actuellement affectés.
-// m = contexte de manifeste { lines, cargo, f, originIdx, destIdx, origin, dest } ; par défaut
-// celui d'« En route », mais une jambe de voyage passe le sien (cf. legSuggestCtx).
-
-
-// Commodités qui pourraient remplir l'espace libre (même origine -> même destination), non chargées.
-// Le calcul vit dans logic.mjs (partagé avec le manifeste optimal, donc éligibilité identique) ;
-// app.js ne fournit que le marché et le résolveur de corrections.
-
-// addableUnits vient de logic.mjs.
-
-// HTML des suggestions de remplissage pour un contexte de manifeste (En route ou jambe de voyage).
-// `addAttrs` = attributs data-* posés sur le bouton d'ajout, propres à l'appelant.
-// Les suggestions de remplissage sont rendues par vues/manifeste.tsx (`<Suggestions>`), pour la
-// carte comme pour une jambe du compagnon. `suggestionsHTML` produisait le MÊME balisage en chaîne
-// et n'a plus lieu d'être : deux fabriques du même bloc auraient fini par diverger.
-
-
-
-// Trouve une commodité par nom OU code (insensible à la casse/espaces). Partagé par les ajouts
-// libres. La résolution vit dans logic.mjs : un code UEX peut désigner deux commodités.
-
-// Ajout LIBRE : n'importe quelle commodité (par nom ou code), même si elle n'est pas vendable à
-// destination — on la charge pour l'écouler ailleurs (ligne « carry-only », marge nulle ici).
-
-
-// Retire une ligne du manifeste (par nom de commodité).
-
-
-// Engager le chargement dans le voyage : le bouton, ou la phrase qui dit pourquoi il n'y est pas.
-// L'état vient de manifestJourneyState (pur, testé) — le rendu ne décide de rien.
-// Le bouton n'existe QUE dans l'état « ajouter », donc la branche REMPLACER d'addToJourney, qui
-// efface un voyage sans prévenir, est inatteignable depuis cette carte.
-// Le bouton « ▶ Ajouter au voyage » et ses phrases sont rendus par vues/manifeste.tsx.
-
-// Dessine le manifeste courant : totaux + lignes (SCU/prix/stock éditables) + suggestions.
-// Tout ce qui suit dépend de l'ÉTAT GLOBAL (le marché, les corrections, le parcours, la
-// composition à la main) ; la mise en forme, elle, vit dans l'îlot.
-
-
-// Recalcule totaux + profit par ligne d'après les SCU saisis, et rafraîchit les suggestions.
-//
-// Elle REPEINT la carte, ce que la version impérative s'interdisait (« elle arracherait le champ
-// qu'on remplit »). Ce n'est plus vrai : le champ SCU est NON CONTRÔLÉ dans l'îlot, donc React
-// garde le même nœud DOM au re-rendu et ne touche ni à sa valeur ni à son curseur. C'est ce qui
-// permet de supprimer les trois écritures en place (`.mprofit`, `.mboxes`, `#manifestTot`) — un
-// nœud possédé par React et muté hors de React, c'est précisément ce que le garde de #113 interdit.
-
-
-// Copie le plan de chargement en texte (pour un 2e écran / des notes).
-
-
-// Le SEUL chemin de sortie de l'app, pour les trois boutons de copie (manifeste, entrepôts,
-// corrections). Un fichier téléchargé était l'autre candidat, écarté parce que les deux issues
-// demandaient une liste à ENVOYER et que ce chemin-ci existait déjà, éprouvé par « ⧉ Copier » et
-// par « Partager » (ADR-006 §3). Il reste techniquement possible : contrairement à ce qu'une
-// première version de l'ADR affirmait, la CSP ne l'interdit pas — la mesure qui le prétendait ne
-// s'est pas reproduite en contre-lecture.
-// `libelle` est le texte à remettre après le retour visuel : chaque bouton a le sien.
-// Un échec de copie DOIT se voir. Il se taisait deux fois (#91) :
-//   - `navigator.clipboard` est absent hors contexte sécurisé (http:// sur un LAN). Le `?.` rendait
-//     alors `undefined`, et le `.then` qui suivait levait une TypeError — la copie ne se faisait pas
-//     et emportait au passage le reste du gestionnaire de clic ;
-//   - un refus de permission tombait dans un `.catch(() => {})` vide, donc sans un mot.
-// Dans les deux cas l'utilisateur repartait en croyant avoir son texte dans le presse-papiers, et
-// collait le contenu précédent. Deux causes, deux messages : « indisponible » et « refusé » ne
-// demandent pas la même chose à qui les lit.
-
-
-
-
-
-
-// ---------- Vue « Chaîne » (multi-sauts A -> B -> C ...) ----------
-
-// buildChainAdjacency vit dans logic.mjs (fonction pure) ; appelée avec MARKET + effVals.
-
-// `chainCardHTML` a été remplacé par vues/chaine.tsx.
-
-
-// ---------- La soute : ce qui est à bord, et ce qu'on l'a payé (ADR-002) ----------
-// Un lot par chargement — la même commodité peut y figurer deux fois à des prix différents.
-// PERSISTÉE ET SANS PÉREMPTION : reprendre le jeu une semaine plus tard avec un vaisseau rangé
-// plein, ce n'est pas une soute périmée, c'est une soute exacte. C'est aussi pour ça qu'effacer le
-// voyage NE VIDE PAS la soute : le parcours est un plan, la soute est du fret réel.
-
-
-// Le REGISTRE des chargements (logic.mjs) : quelle jambe est engagée, et ce qu'elle a pris à quel
-// rayon. Store à part de la soute, et c'est tout le point : la soute se vide par son ✕, par une
-// vente, par la vente implicite du départ — aucun de ces chemins ne rend rien à la station, donc
-// aucun ne décharge la jambe. Le fret peut partir, ce qu'on doit au rayon reste dû.
-// À appeler APRÈS loadSoute : la migration lit les lots pour reconstruire le registre d'une soute
-// écrite avant lui (l'état vivait alors dans la présence des lots, et le stock d'avant dans `avant`).
-
-
-// Charge le manifeste d'une jambe dans la soute, au prix que l'app venait d'afficher. Les lots
-// portent la clé de la jambe : c'est ce qui permet d'annuler un chargement sans deviner.
-// Le point d'achat d'une commodité à un terminal, avec son stock EFFECTIF (corrections comprises)
-// et la date UEX qui sert d'ancre à toute correction locale.
-
-
-// Réécrit la correction de stock d'un point d'achat DEPUIS LE REGISTRE : sa référence, moins tout ce
-// que les jambes encore chargées y prennent. Chargement et annulation posent la même question, et
-// une seule réponse les empêche de diverger — c'est ce qui manquait quand deux jambes achetaient au
-// même point. `prise.ref` sert de repli quand plus aucune jambe ne tient le rayon : on lui rend
-// alors exactement ce qu'il annonçait avant qu'on y touche.
-// Renvoie le solde appliqué, ou null si le point a disparu d'UEX (rien à corriger).
-
-
-
-// L'état « chargée » est PORTÉ par le registre, jamais déduit des lots : la soute se vide par
-// d'autres chemins que « annuler », et aucun ne défait le chargement — le fret est parti, il n'est
-// pas revenu au rayon.
-
-// « Où suis-je ? » — l'étape courante du voyage, ou à défaut le terminal de départ d'« En route ».
-// C'est ce terminal qui fixe le prix d'une vente et qui porte le marqueur « refusé ici ».
-
-// Vend `units` SCU ici. Si le comptoir n'a pas tout pris, le reste est marqué REFUSÉ à cette
-// station : il traversera la vente implicite du départ sans être effacé.
-// `idxFige` : l'index porté par `.hold-sell[data-idx]`, résolu AU RENDU. On vend là où l'utilisateur
-// a lu le prix, pas là où il se trouve à la milliseconde du clic — sinon l'infobulle annonce une
-// station et la vente en encaisse une autre. Repli sur `stationCourante()` pour tout appel qui n'a
-// pas d'affichage derrière lui (venteImplicite, notamment).
-
-
-// Quitter une escale sous-entend qu'on y a fait son affaire : ce qu'elle reprend est vendu.
-// Ce qu'une vente partielle y a laissé porte `refuse` et traverse intact.
-
-
-
-// Le fret déposé à une station : ni vendu, ni perdu — du capital immobilisé qu'on peut oublier.
-
-
-// Même règle que `vendreIci` : on dépose à la station résolue au rendu, celle que le panneau nomme.
-
-
-// Reprendre : le fret déposé remonte à bord avec son prix payé. Aucun contrôle de position — l'app
-// ne sait pas où le vaisseau est RÉELLEMENT, et refuser au motif « tu n'y es pas » bloquerait le
-// geste au moment exact où il est vrai. La station est écrite en toutes lettres sur la ligne :
-// savoir qu'on y est relève de l'utilisateur, pas d'une donnée qu'on n'a pas.
-
-
-// Les entrepôts : le fret déposé, station par station. Masquée tant que rien n'y dort, comme la
-// soute. Elle ne peut PAS vivre dans renderSoute : celle-ci sort dès que la soute est vide, ce qui
-// est justement le cas le plus fréquent quand on vient de tout déposer.
-// Le capital immobilisé est le chiffre qui compte : de l'argent déjà sorti, que plus rien dans
-// l'app ne rappelait — c'était tout le sujet.
-// Les clés de DEPOTS viennent du localStorage : elles sont échappées comme n'importe quelle donnée
-// tierce (cf. e2e/injection.pw.mjs).
-
-
-// La liste de ce qui dort en entrepôt, en texte, dans le presse-papiers : elle ne quittait jamais
-// l'app, et il fallait rouvrir le navigateur qui porte le localStorage pour la relire.
-
-
-// « Où écouler ce qui reste ? » — le détour manuel par la vue Commodités, en un panneau.
-// Le panneau « où écouler » est rendu par vues/soute.tsx (`<OuEcouler>`). Le CALCUL, lui, reste
-// ici : `offloadPlan` a besoin du marché, des filtres et du résolveur de corrections — et il n'est
-// appelé que si le panneau est ouvert, parce qu'il parcourt tous les terminaux à chaque rendu.
-
-// ---------- Déclarer « j'ai ça à bord » : la deuxième entrée de la soute (#55) ----------
-// Le repli que l'ADR-002 réservait (option D) et n'avait jamais câblé. Jusqu'ici rien n'entrait en
-// soute sans une jambe de voyage et son « ✓ chargé » : ni le butin ramassé au sol, ni un vaisseau
-// rangé plein la semaine dernière, ni une cargaison achetée hors du site.
-//
-// Carte SÉPARÉE de #holdCard, et c'est tout le point : celle-ci se masque dès que la soute est
-// vide — un bouton posé dedans serait invisible au moment exact où il sert. Elle vit dans
-// #voyageLeft, que switchView ne touche jamais : les six vues l'ont donc, y compris Trajets et
-// Commodités, qui n'ont aucun rapport avec un voyage.
-
-// Le dessin vit dans `vues/declaration.tsx` ; ici il ne reste que le branchement à l'état.
-//
-// LE GARDE DE FOCUS A DISPARU, et avec lui la relecture des valeurs tapées qui le doublait. Les
-// deux existaient parce qu'un `innerHTML` détruit les champs qu'il réécrit : cette carte est rendue
-// à CHAQUE rafraîchissement — donc à chaque frappe ailleurs et à chaque arrivée du marché — et elle
-// effaçait la saisie en cours. Des champs non contrôlés sous une racine mémorisée rendent le
-// repeint inoffensif : React réutilise leur nœud et n'écrit jamais leur valeur.
-//
-// `synchrone` remplace l'ancien `force`, et ne dit plus du tout la même chose : deux appelants
-// MESURENT ou FOCALISENT juste après le rendu et ne peuvent pas attendre le lot de React. Voir
-// `renderSoute` pour le même contrat sur la carte voisine.
-
-
-function ouvrirDeclaration() {
-  // La vue par défaut ne lit que routes.json : sans le graphe, ni autocomplétion ni résolution du
-  // nom saisi. On l'attend plutôt que d'ouvrir un formulaire inerte.
-  if (!etat.MARKET) { withMarket(ouvrirDeclaration); return; }
-  etat.declarationOuverte = true;
-  // SYNCHRONE, et c'est un contrat : le champ naît avec le formulaire, il n'existe donc pas avant
-  // le rendu. Différé, le `?.` ci-dessous court-circuiterait et le formulaire s'ouvrirait sans
-  // curseur — l'utilisateur taperait dans le vide. Même piège que `.hold-sell-qty` sur la soute.
-  // `flushSync` remplace ici le `{ synchrone: true }` que `peindre()` offrait : le bandeau vit dans
-  // l'arbre, et React rend en différé sauf demande expresse.
-  flushSync(notifier);
-  $("holdAddName")?.focus();
-}
-function fermerDeclaration() { etat.declarationOuverte = false; notifier(); }
-
-// Le geste : cette commodité, ce nombre de SCU, à ce prix. Le prix est FACULTATIF et vaut 0 — du
-// butin n'a rien coûté (ADR-002). Une commodité que le marché ne connaît pas est refusée : l'app ne
-// saurait ni la classer, ni dire où l'écouler, et une ligne muette en soute ne vaut pas mieux que rien.
-
-
-// Poser la position revient à écrire dans le champ de départ d'« En route » : c'est lui que
-// stationCourante() lit sans voyage, et lui que le permalien transporte déjà.
-
-
-// Débarquer le fret n'est pas le remettre en rayon : le registre des chargements n'est pas touché,
-// donc les jambes restent chargées (🔒 « ⬢ à bord ») et leur déduction reste posée. C'est ce qui
-// garde le chemin « annuler » atteignable — le seul qui rende vraiment son stock à la station.
-
-// `synchrone` : la délégation du bouton « vendu » sélectionne le champ de quantité JUSTE APRÈS ce
-// rendu (`…querySelector(".hold-sell-qty")?.select()`). Un rendu React groupé n'aurait pas encore
-// créé le champ, le `?.` court-circuiterait, et le champ s'ouvrirait sans le focus — visible au
-// relevé : sa bordure passait de l'ambre pleine à l'ambre à 30 %.
-
-
-// ---------- Carte 2D du parcours (ADR-001) ----------
-// Le calcul est PUR (journeyMap, logic.ts) et le dessin vit dans `vues/carte.tsx` : ici il ne reste
-// que le branchement à l'état — les globales JOURNEY / MARKET / STARMAP, qu'un îlot ne lit pas.
-
-// Dessine (ou masque) le panneau carte. Appelé par renderJourney, donc à chaque refresh.
-// `renderJourneyMap` est passée dans `vues/carte-vue.tsx`, avec une garde `si="plan"` : elle était
-// rejouée à chaque `refresh()` depuis les huit vues, pour un <aside> que sept d'entre elles
-// n'affichent pas (ADR-012).
-
-
-// ---------- Compagnon de voyage : résumé du parcours (près du vaisseau) ----------
-// `pickJourney` et `syncViewsToJourney` sont passées dans `voyage-actions.ts` : les QUATRE vues qui
-// restent ici les appellent, et aucune ne peut emménager tant qu'elles vivent dans un fichier qui
-// n'exporte rien (ADR-012).
-
-// « Je suis ici » : pose la position courante et recale les vues. Deux chemins y mènent — le fil
-// d'étapes textuel (⦿) et les escales de la carte — et c'est délibéré : la carte n'introduit pas
-// une commande, elle en offre une seconde entrée.
-
-
-// Pré-remplit les contrôles des vues d'après la POSITION COURANTE du parcours.
-// « Pré-rempli » : on pose les défauts, l'utilisateur reste libre de les changer.
-
-
-// `journeyCarriedCommodities`, `figerJambe` et `pinLegsForVolume` sont passées dans
-// `voyage-donnees.ts` : elles ne rendent rien, et le gel doit être appelable depuis une vue de
-// l'arbre (ADR-012). Leurs commentaires les y attendaient déjà, orphelins.
-
-// Engage le manifeste d'« En route » comme nouvelle jambe du voyage (bouton de la carte Manifeste).
-// La garde d'état est REJOUÉE ici : le rendu peut dater d'avant un changement de parcours.
-
-
-// Contexte de manifeste d'une jambe, à la forme attendue par suggestionsFor/manifestRemaining
-// (mêmes suggestions de remplissage qu'« En route »). null si le terminal ou la soute manque.
-
-
-// Actions d'édition d'une jambe (i = index de jambe).
-
-// Saisie en direct : met à jour l'intention, puis repeint la carte AVEC LA GÉNÉRATION FIGÉE.
-//
-// L'ancienne version s'interdisait tout re-rendu — « un renderJourney() à chaque frappe ferait
-// perdre le focus de l'input » — et mettait donc à jour `.jman-profit`, `over-stock` et la boîte de
-// suggestions à la main. C'était vrai d'un `innerHTML`, qui détruit le champ ; ça ne l'est plus
-// d'un rendu React, qui garde le nœud et n'écrit pas dans un champ non contrôlé. Le seul risque
-// serait que la valeur CALCULÉE reprenne la main sous les doigts : `frappe: true` l'empêche en ne
-// bougeant pas la génération (cf. `renderJourney`).
-
-
-// Ajoute une commodité suggérée à une jambe, remplie au max possible.
-
-// « ↺ optimal » lève les deux formes d'intention, l'ajustement manuel comme le gel.
-
-// Ajout LIBRE d'une commodité à une jambe (même non vendable à l'arrivée -> ligne « carry-only »).
-// Même règle qu'« En route » : freeManifestLine (logic.mjs) en est la source unique.
-
-
-// Index du terminal de FIN de parcours (point d'extension), ou null.
-
-// Meilleure jambe (commodité de marge max) entre deux terminaux, ou null si aucun fret rentable.
-// `readFilters()` fait choisir la vente au profit RÉALISABLE et non au prix affiché : sans lui,
-// la jambe proposée peut viser un terminal déjà saturé, qui n'écoulera qu'une poignée de SCU.
-
-// Jambe « à vide » (aucune commodité) entre deux terminaux — pour ajouter un arrêt même sans fret rentable.
-
-// Résout un terminal depuis le texte : libellé exact « Nom — Système », sinon par nom seul.
-// Suggestions d'arrêts : meilleures destinations rentables depuis la fin du parcours (top 4).
-
-// Ajoute un arrêt (terminal) : nouvelle jambe optimale depuis la fin du parcours -> étend.
-
-
-// Démarre un voyage « de zéro » depuis un terminal de départ (sans passer par un trajet ▶).
-// On pose juste le point de départ ; l'utilisateur construit ensuite avec « + Arrêt ».
-
-
-// Réindexe tout ce qui est indexé par le RANG des jambes après une modification du parcours. Les
-// QUATRE porteurs — manifeste édité, 🔒, étiquette `leg` des lots de la soute, entrée du registre
-// des chargements — passent par le même appel pur : les décaler séparément les ferait diverger, et
-// c'est d'en avoir oublié un que venait le double chargement (la jambe renumérotée se croyait vide
-// alors que son fret était à bord).
-
-
-// Retire un arrêt (index de station) et RECONNECTE les voisins (recalcule la jambe A->C).
-
-
-// `frappe` : on est EN TRAIN de taper dans un champ SCU de jambe. La génération ne bouge alors pas,
-// donc les champs non contrôlés gardent leur valeur et leur curseur pendant que profits, totaux et
-// suggestions se recalculent — même mécanique que le manifeste (#117). Tout autre appel la bouge,
-// ce qui remonte les champs : c'est ce que faisait l'ancien `card.innerHTML`.
-
-
-// Récap du voyage (colonne de gauche, sous le vaisseau) : remplit l'espace avec des KPIs utiles.
-
-
-// ---------- La tournée d'écoulement : la huitième vue (ADR-007, #57) ----------
-// « Comment me débarrasser d'une soute que je ne veux plus porter ? » — l'AUTRE question, celle
-// qu'« où écouler » ne pose pas. Les deux vues coexistent parce que les deux classements sont
-// incompatibles, et l'écran doit le dire : sans ça l'app se contredit sous les yeux de
-// l'utilisateur, qui lit deux ordres opposés du même fret.
-
-
-// ---------- Plan de vol : la vue de conclusion (ADR-004) ----------
-// Une CONCLUSION, pas un tableau de bord : on y arrive une fois tout paramétré, pour REGARDER le
-// résultat. Rien n'y est actionnable — pas un bouton de vente, pas un ✕, pas un champ. Les gestes
-// vivent dans les six vues de recherche, et le bandeau (masqué ici, et ici seulement) les y porte.
-
-// Les quatre réglages qui ne FILTRENT pas mais changent le SENS des chiffres (ADR-004 §6). La soute
-// donne la place libre ; l'autoload décide si les profits sont nets ou bruts, et son état n'est
-// autrement lisible que sur une case à cocher — masquée ici. Les taire rendrait la conclusion
-// silencieusement ambiguë : on lirait un profit sans savoir s'il est net.
-
-// Le récapitulatif EN TEXTE (ADR-004 §8). Pas une image : la CSP pose `img-src 'self' https:` sans
-// `data:`, ce qui bloque le procédé habituel (<foreignObject> sérialisé en data: URI). Et le texte
-// se colle partout, se cite ligne par ligne dans un salon, et survit aux thèmes.
-
-
-// Bascule intelligente : si la carte Voyage est bien plus haute que ce qui l'accompagne
-// (voyage long / jambe dépliée), on empile en pleine largeur pour supprimer le grand vide.
-// La carte du parcours ne compte PLUS dans cette mesure : elle a quitté la rangée pour le Plan de
-// vol (ADR-004 §4), où elle occupe toute la largeur. Tant qu'elle était la troisième colonne, il
-// fallait la mesurer — c'était souvent elle qui remplissait le vide. Il ne reste que la colonne de
-// gauche, et le bandeau est de toute façon masqué dans la vue où vit désormais la carte.
-// À appeler APRÈS le rendu des cartes : mesurées avant, elles sont encore masquées (ou à leur
-// taille du rendu précédent) et la rangée s'empilait sur une hauteur qui n'existait plus.
-// Mesure synchrone (getBoundingClientRect force le reflow) -> fiable même onglet non peint.
-
-
-// Bascule entre les vues et rafraîchit la bonne.
-// Regroupe les appels rapprochés en un seul, à la fin de la salve.
-const debounce = (fn, ms = 150) => {
-  let t;
-  return (...a) => { clearTimeout(t); t = setTimeout(() => fn(...a), ms); };
-};
-
-function refresh() {
-  // PLUS AUCUNE BRANCHE DE VUE : les huit vivent dans l'arbre et se réévaluent seules au
-  // `notifier()` de la fin, chacune sous sa propre garde. Ce qui reste ici, c'est le BANDEAU —
-  // visible dans six vues sur huit — et `saveState()`.
-  //
-  // La génération de la carte de chargement bouge ICI et nulle part ailleurs : un cycle complet est
-  // par définition un RECALCUL, donc les champs SCU doivent adopter les nouvelles valeurs. Une
-  // frappe, elle, n'appelle que `notifier()` et laisse la génération tranquille (manifeste-etat.ts).
-  nouvelleGenerationManifeste();
-  nouvelleGenerationVoyage();
-  // La carte Voyage est affichée À CÔTÉ des tableaux, dans toutes les vues : la laisser hors du
-  // cycle de rendu la figeait sur l'état d'avant. Corriger un prix ne mettait donc pas à jour les
-  // bénéfices du voyage — alors qu'une jambe non ajustée est justement, par contrat, branchée sur
-  // le marché et sur les filtres (cf. README). Le coût est celui d'un manifeste par jambe, sur un
-  // parcours qui en compte une poignée ; les champs à saisie libre passent déjà par un debounce.
-  // (le bandeau vit dans l'arbre : `notifier()` suffit)
-  // Effacer le parcours NE VIDE PAS la soute (c'est le contrat, cf. README), et le fret DÉPOSÉ y
-  // survit de même : hors du cycle de rendu, la place libre, « où écouler », le prix du bouton de
-  // vente et les entrepôts restaient figés pendant que les tableaux d'à côté suivaient les filtres.
-  // On appelle les deux rendus DIRECTEMENT plutôt que renderJourney() : la branche « sans voyage »
-  // de celui-ci réécrit l'invite « Nouveau voyage », donc détruirait le champ #journeyStart en
-  // cours de saisie (texte et focus) à chaque frappe faite ailleurs dans l'app. Le coût est nul
-  // quand il n'y a rien : les deux sortent tout de suite en masquant leur carte.
-  saveState();
-  // La propagation vers `etat.ts` (ADR-011). Elle est APPELÉE, jamais déclenchée par une écriture :
-  // des références vives sortent de l'état et sont mutées dehors (`legIntent`), et `logic.ts` mute
-  // `OVERRIDES` en place pendant le rendu — aucun accesseur ne les verrait. Voir l'en-tête d'etat.ts.
-  notifier();
-}
-const refreshDebounced = debounce(refresh);
-
-function switchView(v) {
-  etat.view = v;
-  $("viewRoutes").classList.toggle("active", v === "routes");
-  $("viewLoops").classList.toggle("active", v === "loops");
-  $("viewEnroute").classList.toggle("active", v === "enroute");
-  $("viewChain").classList.toggle("active", v === "chain");
-  $("viewCorrections").classList.toggle("active", v === "corrections");
-  $("viewCommodities").classList.toggle("active", v === "commodities");
-  $("viewPlan").classList.toggle("active", v === "plan");
-  $("viewTour").classList.toggle("active", v === "tour");
-  $("routes").hidden = v !== "routes";
-  $("loops").hidden = v !== "loops";
-  $("enroute").hidden = v !== "enroute";
-  $("enrouteControls").hidden = v !== "enroute";
-  $("chainControls").hidden = v !== "chain";
-  $("chainOut").hidden = v !== "chain";
-  $("correctionsControls").hidden = v !== "corrections";
-  $("corrections").hidden = v !== "corrections";
-  $("commoditiesControls").hidden = v !== "commodities";
-  $("commodities").hidden = v !== "commodities";
-  $("plan").hidden = v !== "plan";
-  $("tourControls").hidden = v !== "tour";
-  $("tour").hidden = v !== "tour";
-  // Les deux blocs jusqu'ici PERMANENTS, que seule la vue de conclusion masque (ADR-004 §4 et §6).
-  // La barre de filtres : on ne change rien au voyage depuis le Plan de vol, l'y laisser ferait
-  // croire le contraire. Le bandeau : ses cartes sont éditables (✕ du parcours, vente en soute) et
-  // c'est tout ce que cette vue n'est pas — le Plan de vol le remplace par un récapitulatif inerte.
-  // Aucune valeur n'est touchée : les deux reviennent intacts au retour dans une vue de recherche.
-  $("controls").hidden = v === "plan";
-  $("shipJourneyRow").hidden = v === "plan";
-  if (v !== "enroute") $("manifest").hidden = true;
-  if (v === "chain" || v === "corrections" || v === "commodities" || v === "plan" || v === "tour") $("empty").hidden = true;
-  refresh();
-}
-
-// Trier est une action à part entière : une souris ne doit pas être la seule façon de la déclencher.
-// Même patron clavier que les valeurs corrigeables (`.editv`) : le clic et Entrée/Espace passent par
-// le MÊME corps, et `tabindex` est posé ici si index.html ne l'a pas fait. Pas de `role="button"`
-// ici, contrairement à `.editv` : il écraserait le rôle `columnheader` du <th>, seul rôle sur lequel
-// `aria-sort` veut dire quelque chose — on perdrait l'annonce de la colonne triée en la corrigeant.
-function sortableHeader(th, apply) {
-  if (!th.hasAttribute("tabindex")) th.tabIndex = 0;
-  th.addEventListener("click", apply);
-  th.addEventListener("keydown", (e) => {
-    if (e.key === "Enter" || e.key === " ") { e.preventDefault(); apply(); } // Espace ne doit pas défiler la page
-  });
-}
-
-function setupSort() {
-  document.querySelectorAll("th[data-sort]").forEach((th) => {
-    sortableHeader(th, () => {
-      const key = th.dataset.sort;
-      if (etat.sortKey === key) etat.sortDir *= -1;
-      else {
-        etat.sortKey = key;
-        etat.sortDir = key === "commodity" ? 1 : -1;
-      }
-      applySortIndicators(); // classes ET aria-sort, pour les deux tables
-      saveState();
-      notifier(); // rendu CIBLÉ : il ne passe pas par `refresh()`, il propage lui-même
-    });
-  });
-}
-
-function setupLoopSort() {
-  document.querySelectorAll("th[data-sort-loop]").forEach((th) => {
-    sortableHeader(th, () => {
-      const key = th.dataset.sortLoop;
-      if (etat.loopSortKey === key) etat.loopSortDir *= -1;
-      else { etat.loopSortKey = key; etat.loopSortDir = -1; }
-      applySortIndicators();
-      saveState();
-      notifier(); // la vue Boucles vit dans l'arbre : la propagation suffit, hors `refresh()`
-    });
-  });
-}
-
-// L'autocomplétion maison, le sélecteur de station et celui des vaisseaux vivent dans
-// `selecteur.js` — avec le garde du champ `#station`, qui n'a de sens qu'à côté de ce qui l'écrit.
-
-// ---------- Persistance & permaliens ----------
-// L'état (filtres, tri, vue, vaisseau) est sauvé dans localStorage ET encodé dans le
-// hash de l'URL, pour reprendre là où on s'est arrêté et partager une vue précise.
-// `alk` = coefficient d'autoload global : partageable, comme tous les réglages. Les relevés PAR
-// STATION, eux, restent locaux — c'est la même frontière que pour les corrections de prix.
-// Positionne l'indicateur ▾/▴ sur la bonne colonne des deux tables.
-// La flèche est un `::after` CSS accroché aux classes : elle n'existe pas pour un lecteur d'écran.
-// `aria-sort` DOUBLE donc les classes (il ne les remplace pas, le CSS s'en sert) sur les seules
-// colonnes triables — le poser sur un <th> décoratif annoncerait une colonne triable qui ne l'est pas.
-function applySortIndicators() {
-  document.querySelectorAll("#routes th, #loops th").forEach((h) => {
-    h.classList.remove("sorted-asc", "sorted-desc");
-    if (h.dataset.sort || h.dataset.sortLoop) h.setAttribute("aria-sort", "none");
-  });
-  if (safeKey(etat.sortKey)) {
-    const th = document.querySelector(`#routes th[data-sort="${etat.sortKey}"]`);
-    if (th) {
-      th.classList.add(etat.sortDir === -1 ? "sorted-desc" : "sorted-asc");
-      th.setAttribute("aria-sort", etat.sortDir === -1 ? "descending" : "ascending");
-    }
-  }
-  if (safeKey(etat.loopSortKey)) {
-    const th = document.querySelector(`#loops th[data-sort-loop="${etat.loopSortKey}"]`);
-    if (th) {
-      th.classList.add(etat.loopSortDir === -1 ? "sorted-desc" : "sorted-asc");
-      th.setAttribute("aria-sort", etat.loopSortDir === -1 ? "descending" : "ascending");
-    }
-  }
-}
-
-
-
-// L'édition en place vit désormais dans `ValeurEditable` (vues/communs.tsx), qui la gère PAR SON
-// ÉTAT. `startEdit` la faisait en mutant le span (`replaceChildren`) depuis une délégation posée
-// sur `document` : c'est la seule situation où les deux modèles se contredisaient vraiment. Les
-// trois comportements qu'elle avait durement acquis — consulter n'écrit rien, Échap ne re-rend
-// pas la vue, le ✎ reste DANS le span — sont portés par le composant, et couverts par
-// `e2e/edition.pw.mjs`.
-
-// `updateOvBadge` est passée dans `corrections-actions.ts`. Elle y reste impérative : le nœud
-// qu'elle touche est dans le RAIL, qu'aucun portail ne possède (ADR-012).
-
-function resetAllOverrides() {
-  if (!ovCount()) return;
-  if (!confirm("Effacer toutes tes corrections locales de prix et de stock ?")) return;
-  resetOverrides();
-  updateOvBadge();
-  refresh();
-}
-
-// Enregistre un relevé de tarif d'autoload pour la station affichée. On mémorise le montant et la
-// quantité observés en plus de `k` : c'est la MESURE qui fait foi, `k` n'en est que la lecture — si
-// la grille change à un patch, un relevé conservé reste réinterprétable.
-function saveStationReading() {
-  const S = indexStationExacte();
-  if (S == null) return;
-  const t = etat.MARKET.terminals[S];
-  const amount = Number($("alAmount").value);
-  const scu = Math.floor(Number($("alScu").value));
-  const k = kFromReading(amount, scu, t.maxBox);
-  if (k == null) { showToast("⚠ Relevé inutilisable — indique le montant payé et la quantité chargée"); return; }
-  // Un montant tapé à côté (un zéro de trop) donne un k d'apparence honnête, qu'on persiste et
-  // qu'on réaffiche « (relevé) » — il se lit alors comme une mesure fiable tout en multipliant les
-  // frais de cette station dans toutes les vues. Hors des bornes plausibles on DEMANDE, on ne
-  // refuse pas : un relevé surprenant reste une mesure, et c'est l'utilisateur qui l'a faite.
-  // Le message montre le montant tel qu'il a été compris et le compare aux deux tarifs connus :
-  // sans ce repère, « k = 1 413 » ne dit pas à quel point c'est absurde.
-  if (!kPlausible(k) && !confirm(
-    `${fmt(amount)} aUEC pour ${fmt(scu)} SCU à ${t.name}, c'est ×${kFmt(k)} le tarif d'Endgame.\n` +
-    `Les deux seules stations mesurées valent ×1 et ×1,4. Un zéro de trop ?\n\nEnregistrer ce relevé quand même ?`
-  )) return;
-  etat.AUTOLOAD_K[alKey(t.name)] = { k, amount, scu };
-  saveAutoloadK();
-  refresh();
-}
-function forgetStationReading(key) { delete etat.AUTOLOAD_K[key]; saveAutoloadK(); refresh(); }
-function resetAllReadings() {
-  if (!Object.keys(etat.AUTOLOAD_K).length) return;
-  if (!confirm("Oublier tous tes relevés de tarif d'autoload ?")) return;
-  etat.AUTOLOAD_K = {};
-  saveAutoloadK();
-  refresh();
-}
-
-// ---------- Vue « Corrections » : liste + édition par station ----------
-// `resolveStation` et la globale `stationSel` sont remplacées par `indexStationExacte()`
-// (marche.ts) : la station se dérive du champ à chaque lecture, au lieu d'être mise en cache par
-// une fonction de rendu. Voir marche.ts pour pourquoi ce n'est pas `resolveStationLabel`.
-
-// `stationFeeHTML` et `autoloadListHTML` sont passées dans `vues/frais-station.tsx` : c'étaient
-// les DEUX DERNIÈRES fonctions de vue rendant des chaînes HTML. Leur garde de signature part avec
-// elles — React ne réécrit pas un champ non contrôlé qu'il réconcilie (ADR-012).
-
-
-
-// Tableau éditable des commodités d'une station (prix/stock à l'achat, prix/demande à la vente).
-// `stationTableHTML` a été remplacé par vues/corrections.tsx.
-// Bandeau COLLANT de la station affichée : sa photo, son nom, sa zone, son code, et ce qu'on y a
-// corrigé. Il remplace la ligne de titre, qui sortait de l'écran au premier coup de molette — après
-// quoi plus rien ne disait quelle station on éditait, au milieu de 92 tuiles.
-// `stationHeroHTML` a été remplacé par vues/corrections.tsx.
-// Retour à la valeur UEX d'un chiffre corrigé. Contrôle DÉDIÉ, posé dans la tuile et hors du
-// `.editv` : celui-ci porte déjà `role="button"`, et y imbriquer un second bouton serait invalide en
-// ARIA — tandis que sortir le `✎` casserait la restauration de startEdit (qui mémorise puis rejoue
-// les childNodes du span) et l'assertion qui la couvre. Il annonce la valeur vers laquelle il
-// ramène : c'est ce que l'ancienne liste plate ne montrait jamais.
-// `retourUEX` a été remplacé par vues/corrections.tsx.
-// La bande : une vignette par station corrigée, la station affichée épinglée en tête. C'est elle qui
-// remplace la liste plate — vingt corrections y tenaient en 900 px, elles tiennent ici en une rangée.
-// `correctionsIndexHTML` a été remplacé par vues/corrections.tsx.
-// Les corrections, en JSON daté, dans le presse-papiers. Du JSON et non du texte libre — celui-ci
-// est fait pour être RELU (cf. relireCorrections), et une correction qu'on ne peut pas dater est
-// une correction qu'on réappliquerait aveuglément des semaines plus tard.
-
-
-
-
-
-
-
-// `renderCorrections`, `tuilesStation` et `groupesCorrections` sont passées dans
-// `vues/corrections-vue.tsx` : la vue vit dans l'arbre (ADR-012).
-
-
-// ---------- Vue « Commodités » ----------
-// `marginTier` est passée dans `logic.ts` sous le nom `palierMarge`, avec le maximum en PARAMÈTRE
-// au lieu de la globale `commMaxMargin` — et donc testable unitairement, ce qu'elle n'était pas.
-
-// Toute la vue Commodités vit désormais dans `vues/commodites-vue.tsx` : la grille, le détail et
-// l'aide y sont rendus par UN composant, dans trois portails (ADR-012 §3). Sont partis d'ici :
-// `paintCommodityDetail`, `syncCommBoardUI`, `setCommBoard`, `setCommSort`, `sortCommodities`,
-// `peindreGrilleCommodites` et `renderCommodities` — plus les cinq caches de rendu du board, qui
-// n'avaient aucun lecteur ailleurs et sont devenus des variables locales du composant.
-
-// Grise le champ soute/budget quand sa contrainte est désactivée.
-function syncToggles() {
-  const cargoOff = !$("useCargo").checked;
-  const budgetOff = !$("useBudget").checked;
-  $("cargo").disabled = cargoOff;
-  $("ship").disabled = cargoOff;
-  $("budget").disabled = budgetOff;
-  // Multi-commodité : remplir la soute n'a pas de sens sans soute bornée -> coche grisée.
-  $("multiCommodity").disabled = cargoOff;
-  $("multiCommodityLabel").classList.toggle("disabled", cargoOff);
-  // Frais d'autoload : le coefficient global n'a de sens que l'interrupteur actif -> champ masqué
-  // sinon (il reste dans l'état, donc dans le lien). La coche, elle, n'est PAS grisée sans soute :
-  // le budget ou le plafond de stock bornent aussi le volume, et un volume borné suffit à facturer.
-  $("alkField").hidden = !$("autoload").checked;
-  // Portée de la liste multi : ne se règle que si la liste multi existe.
-  $("multiModeField").hidden = !$("multiCommodity").checked;
-}
-
-async function init() {
-  // Les deux crochets du chargement : `donnees.ts` sait charger, il ne sait rien des `<datalist>`
-  // ni des messages. `setupEnRoute` DOIT tourner avant chaque rappel de `withMarket` — le déclarer
-  // ici une fois vaut mieux que de le répéter à seize appels.
-  brancher({ apresMarche: setupEnRoute, signalerIndisponible: marketUnavailable });
-  // Le crochet de rendu : c'est ce qui rend `refresh()` joignable depuis un module, et donc depuis
-  // une vue de l'arbre qui porte une ACTION. `notifier()` seule ne suffirait pas — elle ne rejoue
-  // ni la carte du voyage, ni la soute, ni `saveState()`. Voir l'en-tête de `rendu.ts`.
-  brancherRendu({ rafraichir: refresh });
-  // La racine unique. Elle s'abonne à `etat` : à partir d'ici, une vue qui y vit se re-rend
-  // toute seule à chaque `notifier()`, sans que `refresh()` ait à la nommer.
-  monterRacine();
-  setupSort();
-  setupLoopSort();
-  applySortIndicators(); // aria-sort/classes du tri par défaut, sans dépendre des attributs du HTML
-  // Les champs à SAISIE LIBRE sont débouncés : sans ça, chaque caractère relançait un cycle
-  // complet calcul + réécriture de #rows par innerHTML. Mesuré à ~142 ms par frappe sur un CPU
-  // throttlé ×4 (le coût dominant est le relayout de la table, pas le calcul : ~2 ms), soit plus
-  // d'une seconde de thread bloqué pour taper « Laranite », et deux recalculs sur des valeurs
-  // absurdes quand on tape « 696 » dans la soute (6 puis 69 SCU).
-  // Menus et cases à cocher restent IMMÉDIATS : ils n'émettent qu'un seul événement.
-  ["cargo", "budget", "search", "alk"].forEach((id) => $(id).addEventListener("input", refreshDebounced));
-  ["system", "freshness", "sameSystem", "noOutpost", "legalOnly", "capStock", "multiMode"].forEach((id) =>
-    $(id).addEventListener("input", refresh)
-  );
-  // Ces deux-là commandent en plus l'affichage de leur propre sous-réglage (coefficient k, portée
-  // de la liste multi) : ils passent donc par syncToggles avant de recalculer.
-  ["autoload", "multiCommodity"].forEach((id) =>
-    $(id).addEventListener("input", () => { syncToggles(); refresh(); })
-  );
-  ["useCargo", "useBudget"].forEach((id) =>
-    $(id).addEventListener("change", () => {
-      syncToggles();
-      refresh();
-    })
-  );
-  $("viewRoutes").addEventListener("click", () => switchView("routes"));
-  $("viewLoops").addEventListener("click", () => switchView("loops"));
-  $("viewEnroute").addEventListener("click", () => switchView("enroute"));
-  $("viewChain").addEventListener("click", () => switchView("chain"));
-  $("viewCorrections").addEventListener("click", () => switchView("corrections"));
-  $("viewCommodities").addEventListener("click", () => switchView("commodities"));
-  $("viewPlan").addEventListener("click", () => switchView("plan"));
-  $("viewTour").addEventListener("click", () => switchView("tour"));
-  // Terminal à SAISIE LIBRE (datalist, mais rien n'oblige à choisir dedans) : même debounce que
-  // #origin et #indexDepartChaine(), faute de quoi chaque frappe re-rendrait ET réécrirait le hash.
-  $("tourFrom").addEventListener("input", refreshDebounced);
-  $("tourScope").addEventListener("input", refresh); // <select> : un seul événement, immédiat
-  // La marque ramène à TRAJETS, la vue principale — pas au Plan de vol (ADR-004 §5). Un <button>
-  // natif : Entrée et Espace y viennent sans rien ajouter.
-  $("brandHome").addEventListener("click", () => switchView("routes"));
-  // Copie du récapitulatif : écouteur DÉLÉGUÉ sur `#planHead`, que React repeint à chaque rendu.
-  // Posé sur le CONTENEUR et non sur le bouton, il survit donc aux rendus — même règle qu'avant,
-  // pour une autre raison (c'était `renderPlan` qui réécrivait, c'est l'arbre maintenant).
-  $("planHead").addEventListener("click", (e) => {
-    if (e.target.closest("#planCopy")) copierPlan();
-  });
-  $("share").addEventListener("click", copyShareLink);
-  // Contrôles « Commodités » : modes de tri + sélection d'une tuile.
-  $("commSortModes").addEventListener("click", (e) => { const b = e.target.closest("button[data-sort]"); if (b) setCommSort(b.dataset.sort); });
-  $("commBoardModes").addEventListener("click", (e) => { const b = e.target.closest("button[data-board]"); if (b) setCommBoard(b.dataset.board); });
-  // La délégation sur `#commGrid` est partie avec la vue : la tuile est un vrai `<button>`, elle
-  // porte donc son propre `onClick`. La laisser doublerait l'action, ce que seul le compteur de
-  // propagations aurait vu. Les deux au-dessus restent : leurs `<div>` sont du markup d'index.html
-  // qu'aucun portail ne possède, et leurs boutons sont statiques (ADR-012 §2).
-  // Contrôles « En route ». Ces champs de terminal sont eux aussi à SAISIE LIBRE (datalist, mais
-  // rien n'oblige à choisir dans la liste) : même debounce que ci-dessus. Sans lui, chaque frappe
-  // re-rendait la vue ET réécrivait le hash — or WebKit plafonne history.replaceState à 100 appels
-  // par 10 s : taper deux noms de terminal suffisait à le franchir. Les résolveurs restent DANS le
-  // rappel, donc dans le même ordre qu'avant ; les vues de l'arbre les
-  // rejouent de toute façon avant de peindre.
-  // CHANGER DE ROUTE ABANDONNE LA COMPOSITION, et c'est un GESTE — plus une décision du rendu.
-  // `compositionValide` se contente désormais de dire qu'elle ne vaut plus ; c'est ici qu'on
-  // l'efface, parce que c'est ici que l'utilisateur a changé d'avis. La laisser en réserve la
-  // ferait ressurgir au retour sur cette route, longtemps après le geste qui l'avait écrite.
-  const changerDeRoute = () => { oublierCompositionSiRouteChangee(); refresh(); };
-  $("origin").addEventListener("input", debounce(changerDeRoute));
-  $("destSystem").addEventListener("input", changerDeRoute); // <select> : un seul événement, immédiat
-  $("destTerminal").addEventListener("input", debounce(changerDeRoute)); // terminal d'arrivée forcé
-  // Contrôles « Chaîne ».
-  $("chainOrigin").addEventListener("input", debounce(refresh));
-  $("hops").addEventListener("input", refresh);
-  // Contrôles « Corrections » : recherche de station + suppression / reset (délégué).
-  // Ne re-rend QUE si la station résolue a CHANGÉ. Le sélecteur, lui, rend immédiatement au choix :
-  // sans ce garde, le rendu différé du debounce arrivait ~300 ms après et refaisait le même écran
-  // pour rien — en détachant au passage l'éditeur d'un chiffre ouvert entre les deux. Même famille
-  // que #24 : tout re-rendu gratuit de cette vue efface une saisie en cours.
-  $("station").addEventListener("input", debounce(() => { if (stationChangee()) refresh(); }));
-  $("corrections").addEventListener("click", (e) => {
-    // Les relevés d'autoload se testent AVANT les corrections : leur ✕ porte aussi `.corr-del`
-    // (même bouton à l'écran) et tomberait sinon dans la branche qui écrit dans OVERRIDES.
-    const alDel = e.target.closest(".al-del");
-    if (alDel) { forgetStationReading(alDel.dataset.key); return; }
-    // Vignette de la bande : recharge sa station. Écrit le LIBELLÉ CANONIQUE, comme le sélecteur —
-    // la résolution est exacte, et c'est ce libellé-là que le permalien transporte.
-    const tuile = e.target.closest(".stn-tile");
-    if (tuile && !tuile.disabled) {
-      const t = termByName.get(tuile.dataset.terminal);
-      if (t) { $("station").value = stationLabel(t.name, t.system); memoriserStation(); refresh(); saveState(); }
-      return;
-    }
-    // Retour à la valeur UEX d'un chiffre corrigé.
-    const undo = e.target.closest(".scomm-undo");
-    if (undo) {
-      const { c, t: term, s: side, f: field } = undo.dataset;
-      // Rendre son stock à UEX reste un changement de volume : sans ce gel, un voyage déjà planifié
-      // se rebattrait tout seul. Même règle que startEdit et que l'ancien ✕ de la liste plate.
-      if (field === "vol") pinLegsForVolume(c, term, side);
-      const o = etat.OVERRIDES[ovKey(c, term, side)];
-      setOverride(c, term, side, field, null, o ? o.base : 0);
-      updateOvBadge();
-      refresh();
-      return;
-    }
-    // Efface les corrections de la SEULE station affichée (bouton du bandeau).
-    if (e.target.closest("#stnClear")) {
-      const S = indexStationExacte();
-      const nom = S != null ? etat.MARKET.terminals[S].name : null;
-      if (nom) {
-        for (const k of Object.keys(etat.OVERRIDES)) {
-          const [commodity, terminal, side] = k.split("|");
-          if (terminal !== nom) continue;
-          if (etat.OVERRIDES[k].vol != null) pinLegsForVolume(commodity, terminal, side);
-          delete etat.OVERRIDES[k];
-        }
-        saveOverrides(); updateOvBadge(); refresh();
-      }
-      return;
-    }
-    // La branche `.corr-del` générique est partie : elle était MORTE. Les deux seuls producteurs de
-    // cette classe (`vues/frais-station.tsx`) posent aussi `.al-del`, interceptée trois lignes plus
-    // haut avec un `return`. Elle datait de la liste plate de corrections, remplacée par la bande de
-    // vignettes. La garder « par prudence » rouvrirait un chemin d'écriture d'OVERRIDES qui n'existe
-    // plus — et un test asserte au contraire qu'aucun `.corr-item:not(.autoload)` ne subsiste.
-    if (e.target.closest("#alSave")) { saveStationReading(); return; }
-    if (e.target.closest("#resetAllK")) { resetAllReadings(); return; }
-    // Avant « Tout réinitialiser » ici aussi : rien ne doit s'effacer sans qu'on ait pu l'emporter.
-    if (e.target.closest("#exportCorrections")) { copierCorrections(); return; }
-    if (e.target.closest("#resetAll")) resetAllOverrides();
-  });
-  // Validation du relevé d'autoload à la touche Entrée (les deux champs sont dans le même panneau).
-  $("corrections").addEventListener("keydown", (e) => {
-    if ((e.target.id === "alAmount" || e.target.id === "alScu") && e.key === "Enter") { e.preventDefault(); saveStationReading(); }
-  });
-  // Manifeste : ajustement des SCU + ajout (suggéré ou libre) + retrait d'une ligne.
-  $("manifest").addEventListener("input", (e) => {
-    if (e.target.classList.contains("mqty-input")) updateManifestTotals();
-  });
-  $("manifest").addEventListener("click", (e) => {
-    // Ici et pas dans le délégué global du compagnon. La raison d'origine a disparu avec la
-    // délégation `.journey-pick` (elle lisait `pick.closest("table").id`, qui levait depuis une
-    // carte) ; ce qui reste vaut toujours : cet écouteur est posé sur `#manifest`, donc il ne voit
-    // que sa carte, là où le délégué global voit toute la page.
-    if (e.target.closest("#manifestToJourney")) { manifestToJourney(); return; }
-    if (e.target.closest("#copyManifest")) { copyManifest(); return; }
-    if (e.target.closest("#manifestAddBtn")) { addManifestCommodity($("manifestAddInput").value); return; }
-    if (e.target.closest("#manifestReset")) { resetManifeste(); return; }
-    const del = e.target.closest(".mline-del");
-    if (del) { removeManifestLine(del.dataset.name); return; }
-    const add = e.target.closest(".suggest-add");
-    if (add) addSuggestion(add.dataset.name);
-  });
-  $("manifest").addEventListener("keydown", (e) => {
-    if (e.target.id === "manifestAddInput" && e.key === "Enter") { e.preventDefault(); addManifestCommodity(e.target.value); }
-  });
-  // Chargement d'un trajet multi-commodité : déplie/replie le manifeste sous la ligne cliquée.
-  // Le dépliant (#30) n'a plus de délégation : son ouverture est un ÉTAT de `VueTrajetsMulti`, et
-  // sa ligne un frère rendu par le JSX. Celle qui vivait ici injectait une `<tr>` dans `#rows` par
-  // `insertAdjacentHTML` et posait `open` à la main sur un bouton rendu par React — deux écritures
-  // que React ne voyait pas, donc qui survivaient à ses rendus en devenant fausses (#125).
-  //
-  // La garde `isMultiRoutes()` que cette délégation portait disparaît avec elle — elle existait
-  // parce qu'un bouton cliqué après un retour en lignes simples indexait `shownMulti` avec le rang
-  // d'un autre tableau (#25). Un rappel fermé sur SON trajet ne peut pas se tromper de tableau.
-  // `isMultiRoutes` a suivi le même chemin avec le ▶ : plus personne ne demande « quel mode ? »
-  // au moment du clic, puisqu'aucun bouton de CES TABLES n'a plus de rang à interpréter.
-  // Carte du parcours : cliquer une escale déplace « je suis ici », comme le fil d'étapes.
-  $("holdCard").addEventListener("click", (e) => {
-    if (e.target.closest("#holdClear")) { viderSoute(); return; }
-    if (e.target.closest("#holdOffload")) { etat.ecoulerOuvert = !etat.ecoulerOuvert; notifier(); return; }
-    // La quantité ET la station se lisent sur le MÊME conteneur : celui que le rendu a produit.
-    // `dataset.idx` absent -> undefined -> NaN -> repli sur stationCourante() ; jamais 0.
-    const deposer = e.target.closest(".hold-store");
-    if (deposer) { const b = deposer.closest(".hold-sell"); deposerIci(deposer.dataset.name, Number(b.querySelector(".hold-sell-qty").value), Number(b.dataset.idx)); return; }
-    const ouvrir = e.target.closest(".hold-sell-btn");
-    // `flushSync` : le champ de vente doit exister au RETOUR du gestionnaire pour qu'on le
-    // sélectionne. C'est le même contrat que l'ouverture d'une valeur éditable (communs.tsx).
-    if (ouvrir) { etat.venteEnCours = ouvrir.dataset.name; flushSync(notifier); $("holdCard").querySelector(".hold-sell-qty")?.select(); return; }
-    if (e.target.closest(".hold-sell-no")) { etat.venteEnCours = null; notifier(); return; }
-    const ok = e.target.closest(".hold-sell-ok");
-    if (ok) { const b = ok.closest(".hold-sell"); vendreIci(ok.dataset.name, Number(b.querySelector(".hold-sell-qty").value), Number(b.dataset.idx)); return; }
-    const del = e.target.closest(".hold-del");
-    if (del) retirerLot(Number(del.dataset.i));
-  });
-  // Entrée valide la vente, Échap l'annule — même patron que les corrections inline.
-  $("holdCard").addEventListener("keydown", (e) => {
-    if (!e.target.classList.contains("hold-sell-qty")) return;
-    // Entrée doit encaisser à la même station que le bouton ✓ : même index figé, lu sur le conteneur.
-    if (e.key === "Enter") { e.preventDefault(); vendreIci(etat.venteEnCours, Number(e.target.value), Number(e.target.closest(".hold-sell")?.dataset.idx)); }
-    else if (e.key === "Escape") { e.preventDefault(); etat.venteEnCours = null; notifier(); }
-  });
-  // Déclarer « j'ai ça à bord » (#55) : le seul chemin qui fait entrer du fret sans jambe.
-  $("holdDeclare").addEventListener("click", (e) => {
-    if (e.target.closest("#holdAddOpen")) { ouvrirDeclaration(); return; }
-    if (e.target.closest("#holdAddNo")) { fermerDeclaration(); return; }
-    if (e.target.closest("#holdAddOk")) declarerABord();
-  });
-  // Entrée valide depuis n'importe lequel des trois champs, Échap abandonne — même patron que la
-  // vente inline de la soute et que les corrections.
-  $("holdDeclare").addEventListener("keydown", (e) => {
-    if (!e.target.closest(".hold-add")) return;
-    if (e.key === "Enter") { e.preventDefault(); declarerABord(); }
-    else if (e.key === "Escape") { e.preventDefault(); fermerDeclaration(); }
-  });
-  // La position se résout à la frappe, DÉBOUNCÉE comme le champ de départ d'« En route » : c'est le
-  // même champ derrière, et une résolution par caractère repeindrait toute la page à chaque lettre.
-  // On capture la valeur tout de suite : l'événement, lui, sera périmé quand le timer se déclenchera.
-  const poserPositionDifferee = debounce(poserPosition);
-  $("holdDeclare").addEventListener("input", (e) => {
-    if (e.target.id === "holdWhere") poserPositionDifferee(e.target.value);
-  });
-  // Entrepôts : « reprendre » remet le lot en soute. Un seul geste, pas de champ de quantité — on
-  // reprend ce qu'on a laissé ; une reprise partielle se ferait en redéposant. La fonction pure,
-  // elle, accepte déjà des SCU : l'UI pourra suivre sans la toucher.
-  $("depotsCard").addEventListener("click", (e) => {
-    if (e.target.closest("#copyDepots")) { copierEntrepots(); return; }
-    const b = e.target.closest(".depot-take");
-    if (b) reprendreIci(b.dataset.station, b.dataset.name, Number(b.dataset.units));
-  });
-  $("journeyMap").addEventListener("click", (e) => {
-    const a = e.target.closest(".jm-arret");
-    if (a) setJourneyStop(Number(a.dataset.i));
-  });
-  $("journeyMap").addEventListener("keydown", (e) => {
-    const a = e.target.closest(".jm-arret");
-    if (a && (e.key === "Enter" || e.key === " ")) { e.preventDefault(); setJourneyStop(Number(a.dataset.i)); }
-  });
-  // Compagnon de voyage : ✕ efface le parcours, ▶ n'est plus ici.
-  //
-  // Le ▶ était la dernière lecture par rang DES TABLES DE TRAJETS : la délégation lisait
-  // `pick.dataset.row` puis indexait l'un des quatre tableaux `shown*` selon
-  // `pick.closest("table").id`. Un tableau rempli au RENDU et relu au CLIC, avec ce que ça suppose :
-  // que les deux soient d'accord. Chaque îlot ferme désormais son bouton sur SA ligne.
-  //
-  // Le motif SURVIT ailleurs, et il faut le dire pour ne pas croire le dossier clos : la soute
-  // (`retirerLot`), le fil d'étapes, la carte du parcours et surtout l'autocomplétion
-  // (`matches[Number(li.dataset.i)]`) lisent tous un rang posé au rendu. Ce lot ne traite que les
-  // tables de trajets.
-  //
-  // Trois choses partent avec la délégation : le `closest("table").id` (et son TypeError si un ▶
-  // naissait hors table), le `else` fourre-tout qui faisait retomber toute table inconnue sur
-  // `shownRoutes`, et la garde `isMultiRoutes()` qui rattrapait le mode déjà changé au moment du
-  // clic (#25) — un rappel fermé sur SON trajet ne peut pas se tromper de tableau.
-  document.addEventListener("click", (e) => {
-    if (e.target.closest("#journeyClear")) { clearJourney(); return; }
-    // Démarrer un voyage « de zéro » : bouton « Commencer » depuis l'invite.
-    if (e.target.closest("#journeyStartBtn")) { beginJourney($("journeyStart").value); return; }
-    // Ajout d'arrêt : bouton « + Arrêt » ou une suggestion.
-    if (e.target.closest("#journeyAddBtn")) { addStopByTerminal($("journeyAddStop").value); return; }
-    const sug = e.target.closest(".jstop-suggest");
-    if (sug) { addStopByTerminal(sug.dataset.label); return; }
-    // Retirer un arrêt (✕ sur une étape) -> reconnexion des voisins.
-    const del = e.target.closest(".jstep-del");
-    if (del) { removeJourneyStop(Number(del.dataset.i)); return; }
-    // Édition du manifeste d'une jambe : déplier / retirer / ajouter / réinitialiser.
-    const legSug = e.target.closest(".jman-suggest .suggest-add");
-    if (legSug) { addLegSuggestion(Number(legSug.dataset.leg), legSug.dataset.name); return; }
-    const legDel = e.target.closest(".jman-del");
-    if (legDel) { delLegLine(Number(legDel.dataset.leg), legDel.dataset.name); return; }
-    const load = e.target.closest(".jleg-load");
-    if (load) { chargerJambe(Number(load.dataset.leg)); return; } // AVANT .jleg-head : le bouton y vit
-    if (e.target.closest(".jman-reset")) { resetLeg(Number(e.target.closest(".jman-reset").dataset.leg)); return; }
-    const addBtn = e.target.closest(".jman-add-btn");
-    if (addBtn) { addLegLine(Number(addBtn.dataset.leg), addBtn.closest(".jman-add").querySelector(".jman-add-input").value); return; }
-    const head = e.target.closest(".jleg-head");
-    if (head) { toggleLegEditor(Number(head.dataset.leg)); return; }
-    // Parcours interactif : clic sur une étape (⦿) = « je suis ici » -> recale les vues.
-    const step = e.target.closest(".jstep");
-    if (step) setJourneyStop(Number(step.dataset.i));
-  });
-  // L'en-tête d'une jambe est annoncé `role="button"` : Entrée/Espace doivent l'activer comme le clic
-  // (même corps, cf. sortableHeader). On teste `e.target` LUI-MÊME et non `closest()` — modèle
-  // `.editv` plutôt que `.jm-arret` : le bouton « ✓ chargé » vit DANS l'en-tête, et un closest()
-  // déplierait l'éditeur EN PLUS de charger la soute à chaque Entrée sur ce bouton.
-  $("journeyCard").addEventListener("keydown", (e) => {
-    if (!e.target.classList || !e.target.classList.contains("jleg-head")) return;
-    if (e.key !== "Enter" && e.key !== " ") return;
-    e.preventDefault(); // Espace ne doit pas défiler la page
-    toggleLegEditor(Number(e.target.dataset.leg));
-  });
-  // Ajout d'arrêt / de commodité à la touche Entrée.
-  document.addEventListener("keydown", (e) => {
-    if (e.target.id === "journeyStart" && e.key === "Enter") { e.preventDefault(); beginJourney(e.target.value); }
-    else if (e.target.id === "journeyAddStop" && e.key === "Enter") { e.preventDefault(); addStopByTerminal(e.target.value); }
-    else if (e.target.classList && e.target.classList.contains("jman-add-input") && e.key === "Enter") { e.preventDefault(); addLegLine(Number(e.target.dataset.leg), e.target.value); }
-  });
-  // Précharge le marché quand on focus un champ terminal du compagnon -> peuple le datalist.
-  document.addEventListener("focusin", (e) => {
-    if ((e.target.id === "journeyStart" || e.target.id === "journeyAddStop") && !enrouteReady) {
-      if (etat.MARKET) setupEnRoute();
-      else withMarket(() => {});
-    }
-  });
-  // SCU d'une ligne de jambe : suggestions/profit en direct à la frappe, persistance au blur/Entrée.
-  document.addEventListener("input", (e) => {
-    if (e.target.classList && e.target.classList.contains("jman-qty")) liveLegQty(Number(e.target.dataset.leg), Number(e.target.dataset.i), e.target);
-  });
-  document.addEventListener("change", (e) => {
-    if (e.target.classList && e.target.classList.contains("jman-qty")) editLegQty(Number(e.target.dataset.leg), Number(e.target.dataset.i), e.target.value);
-  });
-  // Les deux délégations qui ouvraient l'édition (clic, Entrée/Espace) ont disparu avec
-  // `startEdit` : `ValeurEditable` porte ses propres `onClick`/`onKeyDown`. Le garde `data-react`
-  // qu'elles consultaient n'a plus rien à garder — mais l'attribut RESTE sur le composant, c'est
-  // lui que `e2e/edition.pw.mjs` interroge pour vérifier qu'aucun `.editv` n'échappe à React.
-  // Raccourcis clavier : / (recherche), 1 à 8 (vues). Ignorés pendant la saisie.
-  document.addEventListener("keydown", (e) => {
-    if (e.ctrlKey || e.metaKey || e.altKey) return;
-    const el = document.activeElement;
-    // `role="button"` couvre d'un coup tout ce que l'app rend activable sans être un <button> :
-    // l'en-tête d'une jambe, une escale de la carte, une valeur corrigeable. Sans lui, tabuler
-    // jusqu'à l'un d'eux puis taper « 1 »…« 8 » changeait de vue — l'utilisateur clavier perdait son
-    // contexte au moment précis où il essayait d'agir dessus.
-    if (el && (el.tagName === "INPUT" || el.tagName === "SELECT" || el.tagName === "TEXTAREA" ||
-               el.getAttribute("role") === "button" || el.classList.contains("editv"))) return;
-    if (e.key === "/") { e.preventDefault(); $("search").focus(); }
-    // L'ordre suit celui du rail (#45), et c'est le contrat : le numéro lu sur un bouton EST la
-    // touche qui l'ouvre. Quatre destinations changent de touche — le prix payé une fois pour que
-    // les deux se disent la même chose.
-    else if (e.key === "1") switchView("routes");
-    else if (e.key === "2") switchView("enroute");
-    else if (e.key === "3") switchView("loops");
-    else if (e.key === "4") switchView("chain");
-    else if (e.key === "5") switchView("commodities");
-    else if (e.key === "6") switchView("tour");
-    else if (e.key === "7") switchView("plan");
-    else if (e.key === "8") switchView("corrections");
-  });
-  loadOverrides();
-  loadAutoloadK();
-  loadJourneyEdits();
-  loadJourneyPins();
-  loadManifestEdit();
-  loadSoute();
-  loadChargements(); // après loadSoute : la migration reconstruit le registre depuis les lots
-  loadDepots();
-  updateOvBadge();
-  syncToggles();
-
-  // État à restaurer (URL partagée en priorité, sinon dernière session locale).
-  const saved = loadState();
-
-  try {
-    const [routes, loops, meta] = await Promise.all([
-      fetch("data/routes.json").then((r) => r.json()),
-      fetch("data/loops.json").then((r) => r.json()).catch(() => []),
-      fetch("data/meta.json").then((r) => r.json()).catch(() => null),
-      chargerVaisseaux(),
-    ]);
-    etat.ROUTES = routes;
-    etat.LOOPS = loops;
-
-    // Remplit les filtres système (achat + vente) : #system et la destination « En route ».
-    const systems = [...new Set(routes.flatMap((r) => [r.buy.system, r.sell.system]))].sort();
-    const sel = $("system"), dest = $("destSystem");
-    systems.forEach((s) => {
-      sel.appendChild(new Option(s, s));
-      dest.appendChild(new Option(s, s));
-    });
-
-    if (meta) {
-      const d = new Date(meta.generated_at * 1000);
-      const ageH = (Date.now() / 1000 - meta.generated_at) / 3600;
-      const rel = ageH < 1 ? "il y a moins d'1 h" : ageH < 24 ? `il y a ${Math.round(ageH)} h` : `il y a ${Math.round(ageH / 24)} j`;
-      const stale = ageH > 6; // données rafraîchies chaque heure : au-delà de 6 h, pipeline suspect
-      const tier = stale ? "f-old" : ageH < 3 ? "f-good" : "f-ok"; // couleurs de fraîcheur partagées
-      const exact = d.toLocaleString("fr-FR");
-      // Haut-droite : indicateur de fraîcheur uniquement.
-      $("meta").innerHTML =
-        `<span class="freshness-ind ${tier}" title="Données UEX du ${exact}"><span class="fi-dot"></span>Données ${rel}${stale ? " ⚠" : ""}</span>`;
-      // Bas du rail (« Flux UEX ») : dernière mise à jour + compteurs.
-      const rs = $("railStatus");
-      if (rs) rs.innerHTML =
-        `<div class="rs-updated">Dernière MàJ<br><b>${exact}</b></div>` +
-        `<div class="rs-counts"><b>${meta.routes}</b> routes · <b>${meta.loops ?? etat.LOOPS.length}</b> boucles · <b>${meta.commodities}</b> commodités</div>`;
-      // Version déployée, et le commit qui l'a produite. Sans ça, un rapport de bug n'est
-      // rattachable à rien : on ne sait pas si l'utilisateur regarde le `main` d'il y a dix minutes
-      // ou une coquille servie depuis son cache d'il y a trois semaines. Silencieux si l'estampille
-      // manque — l'amorce versionnée dans data/ ne la porte pas tant qu'un build n'est pas passé,
-      // et « v— » vaudrait moins que rien.
-      const rv = $("railVersion");
-      if (rv && meta.app_version) {
-        rv.textContent = `v${meta.app_version}${meta.commit ? ` · ${meta.commit}` : ""}`;
-        rv.title = `Version déployée${meta.commit ? ` — commit ${meta.commit}` : ""}. À citer dans un rapport de bug.`;
-      }
-    }
-    // Applique l'état restauré une fois le menu système peuplé, puis affiche la bonne vue.
-    // Les trois synchros d'interface restent ici ; le module les appelle DANS le verrou de
-    // restauration, pour qu'aucune ne puisse resauver au milieu.
-    applyState(saved, () => { applySortIndicators(); syncToggles(); refletBoardCommodites(); });
-    montrerCarteVaisseau(); // ré-affiche la carte du vaisseau restauré (image comprise)
-    // Le compagnon de voyage vient d'un permalien, donc de données non fiables. S'il échoue, il ne
-    // doit pas emporter TOUTE l'app dans le catch ci-dessous, qui accuserait alors data/routes.json
-    // — parfaitement chargé — et laisserait l'utilisateur devant une page vide et un message faux.
-    try {
-      rafraichir();
-    } catch (err) {
-      etat.JOURNEY = null;
-      rafraichir();
-      showToast("⚠ Parcours illisible dans le lien — il a été ignoré");
-    }
-    switchView(etat.view);
-  } catch (e) {
-    $("meta").textContent = "Erreur de chargement des données.";
-    $("empty").hidden = false;
-    $("empty").textContent = "Impossible de charger data/routes.json — lance le script de mise à jour.";
-    console.error(e);
-  }
-}
-
-init();
-
-// PWA : installable + consultable hors-ligne (ignoré si non supporté / hors contexte sécurisé).
-if ("serviceWorker" in navigator) {
-  window.addEventListener("load", () => navigator.serviceWorker.register("sw.js").catch(() => {}));
-}
